Refactored code to use pep8 styling guide cleaned requirements.txt
