+++ renson_endura_delta/general_enum.py
