+++ requirements.txt
@@ -1,19 +1,5 @@
-atomicwrites==1.4.0
-attrs==21.2.0
-certifi==2021.5.30
-charset-normalizer==2.0.4
-idna==3.2
-iniconfig==1.1.1
-more-itertools==8.8.0
-packaging==21.0
-pluggy==0.13.1
-py==1.10.0
-pyparsing==2.4.7
 pytest==4.4.1
 pytest-runner==5.3.1
 requests==2.26.0
 requests-mock==1.9.3
-six==1.16.0
-toml==0.10.2
-urllib3==1.26.6
 cachetools==4.2.2
